Combining pss 8 and 9
