+++ ld_script.txt
@@ -137,7 +137,6 @@ SECTIONS {
         src/pokemon_storage_system_6.o(.text);
         src/pokemon_storage_system_7.o(.text);
         src/pokemon_storage_system_8.o(.text);
-        src/pokemon_storage_system_9.o(.text);
         src/pokemon_icon.o(.text);
         src/script_movement.o(.text);
         src/fldeff_cut.o(.text);
